Separated out store stuff into its own module
