+++ src/index.js
@@ -1,27 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import registerServiceWorker from './registerServiceWorker';
-import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
-import thunk from 'redux-thunk';
 
-import rootReducer from './reducers';
-import getRoutes from './routes'
+import getRoutes from './routes';
+import getStores from './stores';
 import './index.css';
 
-const getNewStore = () => {
-  const middlewares = [
-    applyMiddleware(thunk)
-  ];
-  
-  const finalCreateStore = compose(
-    ...middlewares
-  )(createStore);
-
-  return finalCreateStore(rootReducer, {});
-};
-
-ReactDOM.render(<Provider store={getNewStore()}>
+ReactDOM.render(<Provider store={getStores()}>
     {getRoutes()}
   </Provider>, 
   document.getElementById('root')
